feat(tracing): 集成OpenTelemetry实现分布式追踪功能
- 添加core-tracing模块并集成OpenTelemetry自动配置
- 在各服务模块中引入OpenTelemetry依赖和配置
- 配置Jaeger和Loki容器用于链路追踪和日志收集
- 修改OrderDTO中seckillId字段类型从BigInteger改为Long
- 更新MongoDB配置和相关数据访问层代码适配Long类型
- 在各个应用的日志配置中添加OpenTelemetry appender支持
- 更新全局配置文件添加OTLP导出端点配置

+++ goodskill-common/goodskill-core/core-feign/src/main/java/com/goodskill/core/pojo/dto/OrderDTO.java
@@ -4,7 +4,6 @@ import lombok.Data;
 
 import java.io.Serial;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Data
@@ -12,7 +11,7 @@ public class OrderDTO implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    private BigInteger seckillId;
+    private Long seckillId;
 
     private String userPhone;
